Add files via upload
This project is based on mechanism of hand cricket which is a duo game in which one player first bat and creates a target for another one and the one who balls first have to chase the target to win the game. The one who is batting can be out by the bowler before the end of innings just by putting the same number as of the batsmen.